+++ protoc_plugin/lib/src/message_generator.dart
@@ -473,9 +473,7 @@ class MessageGenerator extends ProtobufContainer {
           '@$coreImportPrefix.Deprecated('
           "'See https://github.com/google/protobuf.dart/issues/998.')",
         );
-        out.println(
-          '$classname clone() => $classname()..mergeFromMessage(this);',
-        );
+        out.println('$classname clone() => deepCopy();');
         out.println(
           '@$coreImportPrefix.Deprecated('
           "'See https://github.com/google/protobuf.dart/issues/998.')",