+++ protoc_plugin/CHANGELOG.md
@@ -1,6 +1,10 @@
 ## 23.0.0
 
 * Update generated code for protobuf 5.0.0.
+* Update generated `clone` members to take advantage of faster `deepCopy`
+  implementation in protobuf 5.0.0. ([#742])
+
+[#742]: https://github.com/google/protobuf.dart/pull/742
 
 ## 22.5.0